+++ CopilotKit/packages/react-ui/src/components/chat/Suggestions.tsx
@@ -1,5 +1,4 @@
 import { Suggestion } from "./Suggestion";
-import React from "react";
 import { RenderSuggestionsListProps } from "./props";
 
 export function Suggestions({ suggestions, onSuggestionClick }: RenderSuggestionsListProps) {